split files and rename symbols in schemas
