+++ src/test/extension.test.ts
@@ -1,9 +1,6 @@
 import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
 import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
+
 
 suite('Extension Test Suite', () => {
 	vscode.window.showInformationMessage('Start all tests.');
